Split css template, fixed error
